+++ CHANGELOG.md
@@ -10,6 +10,7 @@ FlatLaf Change Log
   check whether the current look and feel is FlatLaf.
 - Fixed selection background of checkbox in table cell.
 - Fixed jittery submenu rendering on Mac. (issue #10)
+- Button: Make button square if button text is "..." or a single character.
 - ComboBox: Fixed issues with NetBeans `org.openide.awt.ColorComboBox`
   component.
 - Hex color values in `.properties` files now must start with a `#` character.